Change "panic" parameter handling to work more like the kernel
Negative timeout values are treated by the kernel as "reboot
immediately" and 0 is treated as "wait forever".  Emulate this
behaviour in the panic() function.

Treat invalid (non-numeric) values the same as 0, which seems to match
what the kernel does.  Previously we would ignore them completely and
open a shell as normal.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>