init: Always export the resume and noresume variables, as documented
When testing the pending changes to fix unsafe shell usage, I found
that the "resume" variable is unset if resume is disabled.  This
results in an error in the resume script if it's changed to
require that "resume" is set.

Conversely, "noresume" variable is not set if resume is enabled.
Set it to an empty string in that case.  Also, unset it before
exec'ing the real init system.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>