Change get_fstype to be silent on failure
Currently get_fstype prints "unknown" on failure, and some callers
then have to check for this special string.  Change it to return the
empty string on failure.

Most callers check whether it succeeded or not, so this doesn't
make any difference to them.

local_mount_root and hooks/fsck do check for "unknown", so update
them accordingly.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>