README: Summarise what initramfs-tools does and point to manual pages
README hasn't been updated since the initial checkin in 2005, and
the old contents are obsolete: the original goals have either been
achieved or abandoned (we're not  going to build udev against klibc),
and initrds are largely forgotten.

Replace these with a quick summary of what an initramfs is needed for,
how initramfs-tools does it, and a pointer to the manual pages.

Suggested-by: Robie Basak <robie.basak@ubuntu.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>