scripts/local-premount/resume: Report when we are resuming
Resuming can be quite slow on a hard drive, and if plymouth is used
then the output of /bin/resume won't be seen.  Show a message in
plymouth, then hide it if /bin/resume returns.  Roughly based on a
patch by Mario Limonciello.

Thanks: Mario Limonciello <mario.limonciello@dell.com>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>