scripts/functions: Revert change to _log_msg and make callers format-safe
_log_msg used to pass its arguments straight through to printf (but
only if "quiet" is not used), but this was changed by commits
f277309e0b6b and 3650731f3332.  We still need it do that, so that
when callers pass "\\n" it is replaced by a newline.

Change the callers to separate the format and argument strings, fixing
the format-safety problem that the earlier commits were intended to
address.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>