hooks/resume: Disable resume when there are no suitable swap devices
In 0.131 the automatic selection of a resume device was changed to
disable resume if the largest swap device was encrypted with a
ephemeral key.

In 0.132 I changed the automatic selection to also reject zram
devices, but I also changed it to loop over all swap devices since
zram might well be used alongside a conventional swap device that we
could use.  Unfortunately, in case swap devices are available but none
are suitable, this results in selecting the last one.

Closes: #916696
Thanks: Trek <trek00@inbox.ru>
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>