resume: Set plymouth status only if there is a suspend image
We always need to write the resume device name to the kernel, either
(a) to resume, if a suspend image is found, or (b) to enable suspend
to that device later.  We let the kernel distinguish the two cases
itself.

Since we started reporting the attempt to resume through plymouth, if
the text front-end is used (currently the default in Debian) the
message remains on-screen and is confusing in case (b).  (If the
frame-buffer front-end is used, we clear the message.)

I can't think of a message that usefully covers both cases, and if
case (b) it's not really necessary to report anything.  Use fstype to
check for a suspend image before reporting that we're resuming.

Closes: #928736
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>