autopkgtest: Only test hook and boot scripts from specific packages
There are still other packages with hooks and scripts that shellcheck
finds fault with.  This shouldn't cause a test failure for
initramfs-tools itself, but unfortunately that's what happens now:

https://ci.debian.net/data/autopkgtest/testing/amd64/i/initramfs-tools/2780481/log.gz

Replace the Test-Command with a script that tests only hook and boot
scripts installed by specific binary packages.  We can expand this
list as and when other packages are updated.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>