mkinitramfs: Remove armhf dynamic linker hack
This reverts commit da64d1a11ca8680e69d67c192b8d6e60be6411ee, which
was a workaround for the inconsistent choices of interpreter path in
the early days of armhf and the limitations of copy_file at the time.
Now that copy_file automatically replicates symlinks and their
targets, we don't need to do anything special about the dynamic
linker.

An initramfs built for Debian 10 on armhf currently contains:

/lib/arm-linux-gnueabihf/ld-2.28.so           regular file
/lib/arm-linux-gnueabihf/ld-linux-armhf.so.3  link to ld-2.28.so
/lib/arm-linux-gnueabihf/ld-linux.so.3        link to /lib/ld-linux-armhf.so.3
/lib/ld-linux-armhf.so.3                      regular file

There are two *copies* of the dynamic linker as well as two symlinks
to one or the other of them.

After this commit, the initramfs contains:

/lib/arm-linux-gnueabihf/ld-2.28.so           regular file
/lib/arm-linux-gnueabihf/ld-linux-armhf.so.3  link to ld-2.28.so
/lib/ld-linux-armhf.so.3                      link to arm-linux-gnueabihf/ld-2.28.so

The /lib/…/ld-linux.so.3 symlink is no longer created, because Debian
armhf binaries all use /lib/ld-linux-armhf.so.3 as the interpreter
path, but it would be created if we ever switched over.

Signed-off-by: Ben Hutchings <ben@decadent.org.uk>