hook-functions: Search for firmware under /lib/firmware/updates
I like to keep up with the official kernel releases. As a result the
firmware they wind up requiring is newer than what is available in
firmware packages.

I placed the firmware from the git repo in /lib/firmware/updates/
which the kernel searches before /lib/firmware.  Kernel can
find them but update-initramfs cannot.

I have attached an initial patch that solves this. I am currently
using the resultng code and it is working fine for me.

The one thing the patch does not do is allow for a custom location
that can be provided to the kernel to search first.

This is because that's a little more complicated and requires a
decision wrt where to specify this location (ie the config file?).

[bwh: Text above is from AP's bug report with minor edits; I added
 the subject line.]
Closes: #956663
Signed-off-by: Ben Hutchings <ben@decadent.org.uk>