configure_networking(): Wait for specified network device
configure_networking() will issue a `udevadm settle` before trying to
configure an interface. However that's just a "best effort" mechanism
for waiting until all NICs have been discovered. There is no way to
*actually* know that all NICs have been discovered. The USB protocol sets
no time limit on enumeration, for example, and I have a USB NIC that
is consistently discovered after this point.

However, in the case that the user has told us which interface they expect
to be used in the initramfs (via ip=), we can just wait for it specifically.
Bail only if it hasn't appeared within 3 minutes. We can perhaps allow that
timeout to be overridden from the command line in the future.

Closes: #965935
Signed-off-by: dann frazier <dannf@debian.org>