scripts/functions: Refactor device name logic in configure_networking
Make the order of precedence for setting device name explicit.

Split _handle_device_vs_ip into two functions, one to update DEVICE
and one to update IP.  This allows setting the device name in a way
that obviously follows the order of precedence, and also enables
improvements to the device wait loop.

Signed-off-by: Ben Hutchings <benh@debian.org>