mkinitramfs: Generalise kernel compression support check
Since the default compression was changed to zstd and custom kernels
might not enable CONFIG_RD_ZSTD, we check the config for that and fall
back to gzip.

But all of the decompressors are optional, so it makes sense to check
the config in all cases.

This also separates the kernel support check from the compressor
option setup, so there's no need to duplicate the special cases for
gzip.

Signed-off-by: Ben Hutchings <benh@debian.org>