Add hyper-keyboard module, needed to enter LUKS password in Hyper-V
Originally reported in the Kali Linux bug tracker: it's not possible to
type the LUKS password, in case of full disk encryption, in a Gen2
Hyper-V virtual machine. This is due to missing support for keyboard.

The original bug is at https://bugs.kali.org/view.php?id=7846, and I'll
quote the bits that are relevant:

> Because Generation 2 virtual machines in Hyper-V are presented with a
> minimal set of EFI hardware, the kernel module "hyperv_keyboard" must
> be present to interact with the console keyboard. On systems with disk
> encryption, the user will be prompted for the key to decrypt the disk,
> but cannot enter the password because no keyboard driver is present.
>
> [...]
>
> This module *is* present when the grub menu is presented and later
> when the kernel is fully loaded.  FYI, if some mechanism requires a
> mouse or touch, the hid_hyperv module will also be needed.

This issue also affects Debian, as discussed in:
<https://forums.debian.net/viewtopic.php?t=131468>

This issue was also reported in the Ubuntu tracker, back in 2016:
<https://bugs.launchpad.net/ubuntu/+source/initramfs-tools/+bug/1567107>
It was fixed in Ubuntu, but the changes never made it to Debian.

This commit adds the module hyperv-keyboard along with the other
keyboard drivers.

Closes: #1028511
(cherry picked from commit 8f2f49e1f02dcb87e36984bddef113fbb5d04ebf)