hook-functions: auto_add_modules: Add onboard_usb_hub, onboard_usb_dev
The onboard_usb_hub module can handle power regulators for USB devices
it supports. On a MT8186 Chromebook, probing it power-cycles the onboard
USB ports, which causes I/O errors and breaks the boot process if we
have a USB disk mounted as root since it will be forcibly disconnected.

Although it sounds like a kernel bug to power-cycle a hub that is
already in use, other hardware could have onboard hubs that are powered
off before the bootloader starts the kernel, so initializing these hubs
through this module could be necessary for them to boot from USB.

Add the module to the initramfs, so the power-cycle can happen before we
mount disks. It's being renamed to onboard_usb_dev with Linux v6.10, so
add that name to the module list as well.

Signed-off-by: Alper Nebi Yasak <alpernebiyasak@gmail.com>
(cherry picked from commit f495e2905d5f2365a7e22f3c4c058fedf11d2cba)